chore: remove all selectors fom useLabelingCombined
